Resolved confusion with the Ceres cost function and the reprojection error.

+++ lib/camera_model.cpp
@@ -36,7 +36,7 @@ namespace Distortion {
 
     const int minPoints = 3;
 
-    for( int i = 0; i < objectPoints.size(); ++i ) {
+    for( size_t i = 0; i < objectPoints.size(); ++i ) {
       if( objectPoints[i].size() > minPoints && objectPoints[i].size() == imagePoints[i].size() ) {
         result.status[i] = true;
       }
